+++ src/pages/_document.tsx
@@ -1,0 +1,23 @@
+import * as React from 'react';
+import Document, { Html, Head, Main, NextScript, DocumentContext } from 'next/document';
+
+export default class MyDocument extends Document {
+  render() {
+    return (
+      <Html lang="en">
+        <Head>
+          <link rel="dns-prefetch" href="//www.google.co.jp" /> 
+        </Head>
+        <body>
+          <Main />
+          <NextScript />
+        </body>
+      </Html>
+    );
+  }
+}
+
+MyDocument.getInitialProps = async (ctx: DocumentContext) => {
+  const initialProps = await Document.getInitialProps(ctx)
+  return { ...initialProps }
+};
